scripts/requirements.txt
pymongo
dnspython
news-please
cchardet
Pillow
